s3-es/pyproject.toml
[tool.poetry]
name = "s3-es"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.6.1"
python-dotenv = "^0.14.0"
boto3 = "1.14.44"
sentry-sdk = "^0.17.7"
elasticsearch = "^7.9.1"
requests = "^2.24.0"
pandas = "^1.1.3"
fsspec = "^0.8.3"
aiobotocore = "^1.1.2"
s3fs = "^0.5.1"
requests_aws4auth = "^1.0.1"
es_pandas = "^0.0.16"
progressbar2 = "^3.53.1"
python-dateutil = "^2.8.1"
elasticsearch-dsl = "^7.3.0"
backoff = "^1.10.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
pytest = "^6.0.2"
black = { version = "^20.8b1", allow-prereleases = true }

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
